Backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-jwt-extended
werkzeug
python-dotenv
pandas
numpy
torch
transformers
accelerate
scikit-learn
groq
